Ci update (#1)
* Add cmake/rangev3

* Add cmake/gtest

* Simplify CI

* prepare_compiler.sh -> install_compiler.sh

* Add test to CI

* Update README.md

* Remove bootstrap.sh

* Remove modules

* Remove scripts

* Update compiler folder

* Fix github workflows

* Fix toolchain.cmake

* Style fixups

* Test

* Add 04_gmock

* Remove nlohmann tests

+++ src/01_aggregate_reflection/main.cc
@@ -8,7 +8,7 @@ using namespace magic_io;
 struct User {
   std::string name;
   std::string status;
-  uint64_t age;
+  int age;
 };
 
 int main() {
